requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart